task-api/pyproject.toml
[tool.poetry]
name = "textcanvas"
version = "0.1.0"
description = ""
authors = ["Амир Абдрафиков <sequze@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
shared = { path = "../shared", develop = false }
python = "^3.13"
fastapi = "^0.122.0"
pydantic = { extras = ["email"], version = "^2.12.5" }
pydantic-settings = "^2.12.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.44" }
uvicorn = { extras = ["standart"], version = "^0.38.0" }
black = "^25.11.0"
pika = "^1.3.2"
uuid7 = "^0.1.0"
asyncpg = "^0.31.0"
redis = "^7.1.0"
alembic = { extras = ["async"], version = "^1.17.2" }
aio-pika = "^9.5.8"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-mock = "^3.15.1"
aiosmtplib = "^5.0.0"
bcrypt = "^5.0.0"
loguru = "^0.7.3"
slowapi = "^0.1.9"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)"
]
